Move __tests__ folder to top level
