+++ __tests__/store.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import { setActivePinia, createPinia } from 'pinia'
-import { useQueryStore } from '../../store';
+import { useQueryStore } from '../Panel/src/store';
 import fs from 'fs';
 import path from 'path';
