+++ tsconfig.vitest.json
@@ -1,5 +1,6 @@
 {
   "extends": "./tsconfig.app.json",
+  "include": ["env.d.ts", "./__tests__/*"],
   "exclude": [],
   "compilerOptions": {
     "composite": true,
